Add test for peaks2maps.
Also, remove peaks2maps submodule, moving function into meta utils submodule.

Clean up data dir handling as well.
